pyproject.toml
[project]
name = "bq2dynamo"
version = "0.1.0"
description = "BigQuery to DynamoDB data transfer"
requires-python = ">=3.13"
dependencies = [
    "boto3 ~=1.38",
    "google-cloud-bigquery ~=3.33",
    "tenacity ~=9.1",
    "pydantic ~=2.11",
    "pyyaml ~=6.0",
    "msgpack ~=1.1",
    "pybase64 ~=1.4",

    "click ~=8.2",
    "rich-click ~=1.8",

    "icecream",
    "boto3-stubs-lite[dynamodb]",
]
